requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
bleach==3.3.1
boto3==1.20.54
botocore==1.23.54
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-admin-honeypot==1.1.0
django-allauth==0.61.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-summernote==0.8.20.0
gunicorn==22.0.0
idna==3.7
jmespath==0.7.1
oauthlib==3.2.2
packaging==24.0
pathspec==0.10.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.5.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==6.6.0
django-storages==1.14.3